Exceções em EquipamentoServico

+++ back-end/src/main/java/com/basis/srs/servico/EquipamentoServico.java
@@ -1,9 +1,14 @@
 package com.basis.srs.servico;
 
 import com.basis.srs.dominio.Equipamento;
+import com.basis.srs.dominio.Sala;
+import com.basis.srs.dominio.SalaEquipamento;
 import com.basis.srs.repositorio.EquipamentoRepositorio;
+import com.basis.srs.repositorio.SalaRepositorio;
 import com.basis.srs.servico.dto.EquipamentoDTO;
+import com.basis.srs.servico.exception.RegraNegocioException;
 import com.basis.srs.servico.mapper.EquipamentoMapper;
+import com.sun.org.apache.bcel.internal.generic.IF_ACMPEQ;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,9 +22,14 @@ public class EquipamentoServico {
 
     private final EquipamentoRepositorio equipamentoRepositorio;
     private final EquipamentoMapper equipamentoMapper;
+    private final SalaRepositorio salaRepositorio;
 
-    public EquipamentoDTO criar(EquipamentoDTO novoEquipamento){
-
+    public EquipamentoDTO criar(EquipamentoDTO novoEquipamento) {
+        for (Equipamento equipamento : equipamentoRepositorio.findAll()) {
+            if (equipamento.getNome().equals(novoEquipamento.getNome())) {
+                throw new RegraNegocioException("Equipamento já existe");
+            }
+        }
         Equipamento equip = equipamentoRepositorio.save(equipamentoMapper.toEntity(novoEquipamento));
         EquipamentoDTO equipDTO = equipamentoMapper.toDto(equip);
         return equipDTO;
@@ -31,12 +41,31 @@ public class EquipamentoServico {
     }
 
     public EquipamentoDTO buscar(Integer id){
-        EquipamentoDTO dtoEquip = equipamentoMapper.toDto(equipamentoRepositorio.findById(id).orElse(null));
+        EquipamentoDTO dtoEquip = equipamentoMapper.toDto(equipamentoRepositorio.findById(id)
+                .orElseThrow(()->new RegraNegocioException("Usuário não encontrado.")));
         return dtoEquip;
     }
 
     public void deletar(Integer id){
-         equipamentoRepositorio.deleteById(id);
-    }
-    
+        for (Equipamento equipamento:equipamentoRepositorio.findAll()) {
+            if (equipamento.getId()== id && equipamento.getObrigatorio()==0){
+                    for (Sala sala:salaRepositorio.findAll()){
+                        for (SalaEquipamento salaEquipamento : sala.getEquipamentos()) {
+                            if(salaEquipamento.getEquipamento().getId()==id){
+                                throw new RegraNegocioException("Equipamento está sendo utilizado em uma sala.");
+                        }else{
+                                equipamentoRepositorio.deleteById(id);
+                            }
+                        }
+                    }
+
+            }else {
+                throw new RegraNegocioException("Equipamento é obrigatório");
+                }
+            }
+        }
+
 }
+
+
+
